# internal/models/user.go
package models

import (
	"log"
	"strings"
	"time"

	"golang.org/x/crypto/bcrypt"
	"gorm.io/gorm"
)

type User struct {
	ID        uint      `json:"id" gorm:"primaryKey"`
	FirstName string    `json:"first_name" validate:"required,min=2,max=50"`
	LastName  string    `json:"last_name" validate:"required,min=2,max=50"`
	Age       int       `json:"age" validate:"required,min=1,max=150"`
	Email     string    `json:"email" gorm:"unique" validate:"required,email"`
	Password  string    `json:"password,omitempty" validate:"required,min=6"`
	Posts     []Post    `json:"posts,omitempty" gorm:"foreignKey:UserID;constraint:OnDelete:CASCADE"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

func (u *User) BeforeSave(tx *gorm.DB) error {
	if u.Password != "" {
		// Check if the password is already hashed
		if !strings.HasPrefix(u.Password, "$2a$") {
			hashedPassword, err := bcrypt.GenerateFromPassword([]byte(u.Password), bcrypt.DefaultCost)
			if err != nil {
				return err
			}
			u.Password = string(hashedPassword)
		}
	}
	return nil
}

func (u *User) ComparePassword(password string) error {
	log.Printf("Comparing passwords - Stored hash: %s, Input password: %s", u.Password, password)
	err := bcrypt.CompareHashAndPassword([]byte(u.Password), []byte(password))
	if err != nil {
		log.Printf("Password comparison error: %v", err)
	}
	return err
}

type UserResponse struct {
	ID        uint      `json:"id"`
	FirstName string    `json:"first_name"`
	LastName  string    `json:"last_name"`
	Age       int       `json:"age"`
	Email     string    `json:"email"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}